fix: Transplant correct engine core into modular architecture
